コメントとか足した/構文エラーのときの表示改善
